Refactor gh link and add keys.
