+++ frontend/src/app/[locale]/roadmap/RoadmapWhatWereWorkingOn.tsx
@@ -1,7 +1,7 @@
 import { useMessages, useTranslations } from "next-intl";
 import { Grid, GridContainer } from "@trussworks/react-uswds";
 
-import { GithubIssueLink, gitHubLinkForIssue } from "../process/ProcessNext";
+import GithubIssueLink, { gitHubLinkForIssue } from "src/components/GithubLink";
 import RoadmapPageSection from "./RoadmapPageSection";
 
 export default function RoadmapWhatWereWorkingOn() {
@@ -15,9 +15,12 @@ export default function RoadmapWhatWereWorkingOn() {
       sectionContent={
         <GridContainer className="padding-0">
           {contentItems.map((i, j) => (
-            <Grid row>
+            <Grid row key={`roadmap-what-were-working-on-${j}`}>
               {i.map((k, l) => (
-                <Grid tablet={{ col: 6 }}>
+                <Grid
+                  key={`roadmap-what-were-working-on-${j}-${l}`}
+                  tablet={{ col: 6 }}
+                >
                   <h3 style={{ fontSize: 18 }}>
                     {t(`contentItems.${j}.${l}.title`)}
                   </h3>
